backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
authlib==1.2.1
email-validator>=2.3.0
reportlab==4.0.7
PyPDF2==3.0.1
cryptography==41.0.0
gmpy2>=2.2.0

# Новые зависимости для исправлений
aiosmtplib>=2.0.0
redis>=5.0.0
